# Conflicts:
#	CMakeLists.txt
#	Common/ac/characterinfo.cpp
#	Common/ac/characterinfo.h
#	Common/core/def_version.h
#	Common/game/interactions.cpp
#	Common/game/main_game_file.cpp
#	Common/game/room_file.cpp
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Native/agsnative.cpp
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Engine/ac/character.cpp
#	Engine/ac/characterinfo_engine.cpp
#	Engine/ac/dynobj/cc_staticarray.h
#	Engine/ac/gamestate.cpp
#	Engine/ac/gamestate.h
#	Engine/ac/global_character.cpp
#	Engine/ac/route_finder_impl_legacy.cpp
#	Engine/ac/string.cpp
#	Engine/game/game_init.cpp
#	Engine/game/savegame_internal.h
#	Engine/game/savegame_v321.cpp
#	Engine/main/game_run.cpp
#	Engine/platform/android/acpland.cpp
#	Engine/platform/emscripten/acpemscripten.cpp
#	Engine/platform/ios/acplios.cpp
#	Engine/platform/osx/acplmac.cpp
#	version.json

+++ Engine/platform/windows/win_ex_handling.cpp
@@ -87,8 +87,8 @@ int malloc_fail_handler(size_t amountwanted)
 {
     char *cur = tempmsg;
     char const* const end = tempmsg + sizeof tempmsg;
-    const int MB = 1024*1024;
-    const int KB = 1024;
+    const size_t MB = 1024*1024;
+    const size_t KB = 1024;
 #ifdef USE_CUSTOM_EXCEPTION_HANDLER
     CreateMiniDump(NULL);
 #endif
@@ -100,7 +100,7 @@ int malloc_fail_handler(size_t amountwanted)
     const size_t total_lockspr = spriteset.GetLockedSize();
     const size_t total_extspr = spriteset.GetExternalSize();
     const size_t max_normspr = spriteset.GetMaxCacheSize();
-    const unsigned norm_spr_filled = max_normspr > 0 ?  (uint64_t) (total_normspr * 100 / max_normspr) : 0;
+    const unsigned norm_spr_filled = max_normspr > 0 ?  ((uint64_t)total_normspr * 100 / max_normspr) : 0;
 
     cur += snprintf(cur, end-cur, "Sprite cache KB: %zu / %zu (%u%%), locked: %zu, ext: %zu\n\n",
                     total_normspr/KB, max_normspr/KB, norm_spr_filled, total_lockspr/KB, total_extspr/KB);
